refactor(schema-utils): move MembershipValidator from tenant-api to schema-utils
