+++ packages/schema-utils/src/validation/index.ts
@@ -3,3 +3,4 @@ export * from './ModelValidator'
 export * from './ValidationValidator'
 export * from './errors'
 export * from './SchemaValidator'
+export * from './MembershipValidator'
